More code cleanup.

+++ src/core/EventHandler.cpp
@@ -13,7 +13,7 @@ namespace Vulture2D {
     }
 
     void EventHandler::removeListener(int id) {
-        for (int i = 0; i < listeners.size(); i++) {
+        for (size_t i = 0; i < listeners.size(); i++) {
             if (listeners.at(i).id == id) {
                 listeners.erase(listeners.begin() + i);
             }
